Python fold matching include up to two empty lines
PEP8 suggests that function and class definitions should be followed by two
newlines. The folding option reflects this by including these newlines into the
folds.

One exception are newlines before end of the file. These are not folded,
because I try to avoid empty lines at the end of the file